Ucb 5

+++ src/main/java/ucb/internship/backend/services/InternshipService.java
@@ -1,14 +1,12 @@
 package ucb.internship.backend.services;
+import ucb.internship.backend.dtos.InternshipDto;
+import ucb.internship.backend.models.Internship;
 
 import java.sql.Date;
 import java.util.List;
 
 import ucb.internship.backend.dtos.*;
-import ucb.internship.backend.models.Internship;
-import java.sql.Timestamp;
-import java.util.Collection;
 import org.springframework.data.domain.Page;
-import ucb.internship.backend.dtos.InternshipDto;
 
 public interface InternshipService {
     String createInternship(InternshipDto internshipDTO);
@@ -23,7 +21,7 @@ public interface InternshipService {
 
     List<ActiveInternshipDto> getActiveInternshipsByInstitutionId(Long id);
 
-    List<ApplicantDto> getApplicantsByInternshipId(Integer id);
+    List<ApplicantDto> getApplicantsByInternshipId(Long id);
 
     Page<InternshipListDto> filterInternships(String city , Date startingDate, Date endingDate, String major, Integer page, Integer size);
 }
